requirements.txt
numpy==1.26.4

# FastAPI and dependencies
fastapi==0.78.0
uvicorn==0.18.2
pydantic==1.10.11
python-multipart==0.0.5
anyio==3.6.1

# Dependency Injection
dependency-injector==4.39.1

# Configuration management
omegaconf==2.3.0
PyYAML==6.0

# Testing and linting
pytest==7.4.0
pytest-httpserver==1.0.8
pytest-cov==4.1.0
flake8==7.1.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==23.12.2
flake8-commas==2.1.0
flake8-comprehensions==3.15.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-quotes==3.4.0
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
wemake-python-styleguide==0.18.0
darglint==1.8.1
mccabe==0.7.0
pep8-naming==0.13.3
pycodestyle==2.12.0
pyflakes==3.2.0

# Model 1: Detection
# Essential packages with specified versions
onnx==1.12.0

# YOLOv5 dependencies
opencv-python==4.10.0.82
pillow>=10.3.0
ultralytics>=8.2.34

# YOLO inference
onnxruntime>=1.20.1
coloredlogs>=15.0.1
humanfriendly>=10.0

# Model 2: OCR
albumentations==1.3.1